+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Fix problem letter tag in multipart questions
+* Add `lifestage-context` note to `nursing-internal`
+* Add section titles links to `summary` in `nursing-internal`
 * Change EOC titles in `nursing-internal`
 
 ## [v2.5.0] - 2023-11-03